requirements.txt
Django
Celery
django-celery-results
gpiozero
RPi.GPIO
tzlocal
python-dateutil